Harden preprocessing validation
- Add a final Great Expectations validation step for model-ready
preprocessing artifacts, including selected-feature schema checks,
dtype/null/finite-value checks, binary label checks, and X/y index
alignment.
- Tighten raw-data validation to enforce exact input dtypes and preserve
float dtype for empty selected-feature frames.
- Update the Kedro pipeline and catalog to route intermediate candidate
artifacts through the new validation stage before persistence.

+++ src/mlops_project/pipelines/preprocessing/nodes.py
@@ -286,7 +286,7 @@ def _transform_split(transformer: Any, split: pd.DataFrame) -> pd.DataFrame:
 def _select_split_columns(split: pd.DataFrame, selected_features: list[str]) -> pd.DataFrame:
     """Return one split restricted to the ordered selected feature list."""
     if split.empty:
-        return pd.DataFrame(columns=selected_features, index=split.index)
+        return pd.DataFrame(columns=selected_features, index=split.index, dtype=float)
 
     return split.loc[:, selected_features].copy()
